requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.0.5
async-lru==2.0.4
attrs==23.1.0
Babel==2.11.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==4.1.0
blis==0.7.9
blosc2==3.10.2
bokeh==3.4.1
Bottleneck==1.3.7
branca==0.6.0
Brotli==1.0.9
catalogue==2.0.10
certifi==2025.10.5
cffi==1.16.0
cftime==1.6.5
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.1
confection==0.1.4
contourpy==1.2.0
cookiecutter==2.6.0
cookiecutter-data-science==2.3.0
coverage==7.6.4
cycler==0.11.0
cymem==2.0.6
dask==2025.10.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dotenv==0.9.9
einops==0.8.1
empath==0.89
en-core-web-sm==3.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.20.0
fiona==1.9.5
folium==0.14.0
fontawesomefree==6.6.0
fonttools==4.51.0
frozenlist==1.8.0
fsspec==2025.9.0
funcy==2.0
Fuzzy==1.2.2
fuzzywuzzy==0.18.0
GDAL==3.6.2
gensim==4.3.3
geopandas==0.14.2
h11==0.14.0
h5py==3.15.1
httpcore==1.0.2
httpx==0.27.0
idna==3.7
importlib_metadata==8.7.0
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.27.0
ipython-genutils==0.2.0
jedi==0.19.1
jellyfish==1.1.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.2.5
jupyterlab-pygments==0.1.2
jupyterlab_server==2.27.3
kiwisolver==1.4.4
langcodes==3.3.0
lib==4.0.0
lightning-utilities==0.15.2
locket==1.0.0
loguru==0.7.3
mapclassify==2.5.0
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.0
Metaphone==0.6
mistune==2.0.4
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.7
nb-conda-kernels==2.3.1
nbclassic==1.1.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
ndindex==1.10.0
nest-asyncio==1.6.0
netCDF4==1.7.3
networkx==3.3
nltk==3.9.1
notebook==6.5.7
notebook_shim==0.2.3
numexpr==2.14.1
numpy==1.26.4
openpyxl==3.1.5
overrides==7.4.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.2
patsy==0.5.6
pexpect==4.8.0
phonetics==1.0.5
pillow==10.4.0
pip==24.2
platformdirs==3.10.0
plotly==5.24.1
pluggy==1.5.0
pooch==1.8.2
pprintpp==0.4.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.43
propcache==0.4.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pvlib==0.13.1
py-cpuinfo==9.0.0
pycountry==24.6.1
pycountry-convert==0.7.2
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.15.1
pyLDAvis==3.4.1
pyparsing==3.0.9
pyproj==3.6.1
PySocks==1.7.1
pytest==8.3.3
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.7
python-Levenshtein==0.12.2
python-louvain==0.16
python-slugify==8.0.4
pytorch-lightning==2.5.5
pytz==2024.1
pywaffle==1.1.1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.30.2
regex==2024.9.11
repoze.lru==0.7
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.10.6
Rtree==1.0.1
scikit-learn==1.1.3
scipy==1.10.0
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==72.1.0
shapely==2.0.5
shellingham==1.5.0
six==1.16.0
sklearn-pandas==2.2.0
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.8
stack-data==0.2.0
statsmodels==0.14.2
sympy==1.14.0
tables==3.10.2
tenacity==9.0.0
terminado==0.17.1
text-unidecode==1.3
textblob==0.18.0.post0
thinc==8.2.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.2.1
tomlkit==0.13.3
toolz==1.1.0
torch==2.8.0
torch-geometric==2.6.1
torch_scatter==2.1.2
torch_sparse==0.6.18
torch_spatiotemporal==0.9.5
torchmetrics==1.8.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
tsl==2.0
typer==0.9.0
types-python-dateutil==2.9.0.20250822
typing_extensions==4.11.0
tzdata==2023.3
unicodedata2==15.1.0
urllib3==2.2.2
vaderSentiment==3.3.2
wasabi==0.9.1
wcwidth==0.2.5
weasel==0.3.4
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
wordcloud==1.9.4
xarray==2025.10.1
xlrd==2.0.1
xyzservices==2022.9.0
yarl==1.22.0
zipp==3.23.0
